B06-CRUD-Practice/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
python-dotenv==0.14.0
uuid==1.30
Werkzeug==1.0.1
